requirements.txt
flet==0.24.1
mysql-connector-python==9.0.0
bcrypt==4.2.0
python-dotenv==1.0.0

pytest==7.4.0
